% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CALINE3-package.R, R/libcaline3-CAL3RXL.R, R/libcaline3-CAL3RXM.R
\docType{package}
\name{CALINE3}
\alias{CAL3RXL}
\alias{CAL3RXM}
\alias{CALINE3}
\alias{CALINE3-package}
\title{CALINE3}
\usage{
CAL3RXL(XR, YR, ZR, XL1, YL1, XL2, YL2, WL, HL, NTYP, VPHL, EFL, UM, BRGM,
  CLASM, MIXHM, ATIM, Z0, VS, VD, LXR, .coerce = TRUE)

CAL3RXM(XR, YR, ZR, XL1, YL1, XL2, YL2, WL, HL, NTYP, VPHL, EFL, UM, BRGM,
  CLASM, MIXHM, ATIM, Z0, VS, VD, LXR, .coerce = TRUE)
}
\arguments{
\item{XR}{x-coordinates of the receptors}

\item{YR}{y-coordinates of the receptors}

\item{ZR}{z-coordinates of the receptors (height above ground level, usually 1.8m)}

\item{XL1}{starting x-coordinates of the links}

\item{YL1}{starting y-coordinates of the links}

\item{XL2}{ending x-coordinates of the links}

\item{YL2}{ending y-coordinates of the links}

\item{WL}{widths of the links}

\item{HL}{heights of the links (above ground level)}

\item{NTYP}{link classifications (1=at grade, 2=bridge, 3=fill, 4=depressed)}

\item{VPHL}{link-level traffic volumes, in vehicles per hour}

\item{EFL}{link-level emission factors, in grams per vehicle-mile per hour}

\item{UM}{wind speeds, in meters per second (not less than 1.0)}

\item{BRGM}{wind bearings, in degrees (direction wind is blowing from)}

\item{CLASM}{stability classes (1, 2, 3, 4, 5, or 6)}

\item{MIXHM}{mixing heights, in meters (over 1000 skips mixing height calculations)}

\item{ATIM}{averaging time, in minutes (usually 60)}

\item{Z0}{surface roughness, in centimeters}

\item{VS}{settling velocity, in cm/sec}

\item{VD}{deposition velocity, in cm/sec}

\item{LXR}{whether to process (or skip) a given receptor-link pair}

\item{.coerce}{force arguments to be cast to correct type}
}
\value{
CAL3RXL returns a matrix of concentrations of size NR x NL,
        where NR is the number of receptors and NL is the number of links.

CAL3RXM returns a matrix of concentrations of size NR x NM,
        where NR is the number of receptors and NM is the number of meteorological conditions
}
\description{
CALINE3 provides an interface to a Fortran implementation of CALINE3.

CAL3RXL predicts the incremental contribution from each link
to the concentration at each receptor under a single set of meteorological conditions.

Given a sample of representative meteorological conditions, CALINE3_RECEPTOR_TOTALS
predicts cumulative concentrations at each receptor (from incremental concentrations
contributed by each link)
}
\details{
All coordinates are in meters unless otherwise specified. By default, predicted
concentrations are returned in units of grams per cubic meter (µg/m^3).
}
\examples{
CAL3RXL(
  XR = 30., YR = 0., ZR = 1.8,
  XL1 = 0., YL1 = -5000., XL2 = 0., YL2 = 5000.,
  WL = 30., HL = 0., NTYP = 1, VPHL = 7500., EFL = 30.,
  UM = 1.0, BRGM = 270., CLASM = 6, MIXHM = 1000.,
  ATIM = 60., Z0 = 10., VS = 0., VD = 0.)
CAL3RXM(
  XR = 30., YR = 0., ZR = 1.8,
  XL1 = 0., YL1 = -5000., XL2 = 0., YL2 = 5000.,
  WL = 30., HL = 0., NTYP = 1, VPHL = 7500., EFL = 30.,
  UM = 1.0, BRGM = 270., CLASM = 6, MIXHM = 1000.,
  ATIM = 60., Z0 = 10., VS = 0., VD = 0.)
}
\author{
David Holstius \email{dholstius@baaqmd.gov}
}
\references{
Benson, P. (1979) CALINE-3: A versatile dispersion model for predicting air pollution levels near highways and urban streets. Federal Highway Authority report FHWA/CA/TL-79/23, California DOT, Sacramento (1979).
}
\keyword{CALINE}
\keyword{CALINE3}
\keyword{disperson}
\keyword{package}
\keyword{traffic}

