% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{CALINE3}
\alias{CALINE3}
\alias{CALINE3-package}
\alias{CALINE3_MATRIX}
\alias{CALINE3_RECEPTOR_TOTALS}
\title{CALINE3}
\usage{
CALINE3_RECEPTOR_TOTALS(XR, YR, ZR, XL1, YL1, XL2, YL2, WL, HL, TYP, VPHL, EFL,
  UM, BRGM, CLASM, MIXHM, ATIM, Z0, VS, VD)

CALINE3_MATRIX(XR, YR, ZR, XL1, YL1, XL2, YL2, WL, HL, NTYP, VPHL, EFL, U, BRG,
  CLAS, MIXH, ATIM, Z0, VS, VD, .coerce = TRUE)
}
\arguments{
\item{XR}{x-coordinates of the receptors}

\item{YR}{y-coordinates of the receptors}

\item{ZR}{z-coordinates of the receptors (height above ground level, usually 1.8m)}

\item{XL1}{starting x-coordinates of the links}

\item{YL1}{starting y-coordinates of the links}

\item{XL2}{ending x-coordinates of the links}

\item{YL2}{ending y-coordinates of the links}

\item{WL}{widths of the links}

\item{HL}{heights of the links (above ground level)}

\item{NTYP}{link classifications (1=at grade, 2=bridge, 3=fill, 4=depressed)}

\item{VPHL}{link-level traffic volumes, in vehicles per hour}

\item{EFL}{link-level emission factors, in grams per vehicle-mile per hour}

\item{U}{wind speeds, in meters per second (not less than 1.0)}

\item{BRG}{wind bearings, in degrees (direction wind is blowing from)}

\item{CLAS}{stability classes (1, 2, 3, 4, 5, or 6)}

\item{MIXH}{mixing heights, in meters (over 1000 skips mixing height calculations)}

\item{ATIM}{averaging time, in minutes (usually 60)}

\item{Z0}{surface roughness, in centimeters}

\item{VS}{settling velocity, in cm/sec}

\item{VD}{deposition velocity, in cm/sec}

\item{XR}{x-coordinates of the receptors}

\item{YR}{y-coordinates of the receptors}

\item{ZR}{z-coordinates of the receptors (height above ground level, usually 1.8m)}

\item{XL1}{starting x-coordinates of the links}

\item{YL1}{starting y-coordinates of the links}

\item{XL2}{ending x-coordinates of the links}

\item{YL2}{ending y-coordinates of the links}

\item{WL}{widths of the links}

\item{HL}{heights of the links (above ground level)}

\item{NTYP}{link classifications (1=at grade, 2=bridge, 3=fill, 4=depressed)}

\item{VPHL}{link-level traffic volumes, in vehicles per hour}

\item{EFL}{link-level emission factors, in grams per vehicle-mile per hour}

\item{U}{wind speeds, in meters per second (not less than 1.0)}

\item{BRG}{wind bearings, in degrees (direction wind is blowing from)}

\item{CLAS}{stability classes (1, 2, 3, 4, 5, or 6)}

\item{MIXH}{mixing heights, in meters (over 1000 skips mixing height calculations)}

\item{ATIM}{averaging time, in minutes (usually 60)}

\item{Z0}{surface roughness, in centimeters}

\item{VS}{settling velocity, in cm/sec}

\item{VD}{deposition velocity, in cm/sec}

\item{.coerce}{force arguments to be cast to correct type}
}
\value{
NR x NM matrix of concentrations, where NR is the number of receptors
        and NM is the number of meteorological conditions

NR x NL matrix of concentrations, where NR is the number of receptors
        and NL is the number of links
}
\description{
CALINE3 provides an interface to a Fortran implementation of CALINE3.

For each receptor, predict cumulative concentrations (contributed by all links)
under a number of different meteorological conditions.

For each receptor, predict the incremental contribution from each link
under a single set of meteorological conditions.
}
\details{
All coordinates are in meters unless otherwise specified.

All coordinates are in meters unless otherwise specified.
}
\author{
David Holstius \email{david.holstius@berkeley.edu}
}
\references{
Benson, P. (1979) CALINE-3: A versatile dispersion model for predicting air pollution levels near highways and urban streets. Federal Highway Authority report FHWA/CA/TL-79/23, California DOT, Sacramento (1979).
}
\keyword{CALINE}
\keyword{CALINE3}
\keyword{disperson}
\keyword{package}
\keyword{traffic}

