\docType{package}
\name{CALINE3}
\alias{CALINE3}
\alias{CALINE3-package}
\title{CALINE3}
\description{
  CALINE3 provides an interface to a Fortran implementation
  of CALINE3.
}
\author{
  David Holstius \email{david.holstius@berkeley.edu}
}
\references{
  Benson, P. (1979) CALINE-3: A versatile dispersion model
  for predicting air pollution levels near highways and
  urban streets. Federal Highway Authority report
  FHWA/CA/TL-79/23, California DOT, Sacramento (1979).
}
\keyword{CALINE}
\keyword{CALINE3}
\keyword{disperson}
\keyword{package}
\keyword{traffic}

